+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# v0.4.X
+# 0.X.Y
 
+- Add CronJob support (for kubernetes >=1.24.0-beta.0)
 - Hide confusing log message about `--kubeconfig` not specified
 
 # v0.4.0